requirements.txt
requests
beautifulsoup4
psycopg2-binary
streamlit
pandas
pgvector
Faker
numpy
pyyaml
jinja2
weasyprint
supabase<2.0.0
langchain==0.2.16
langchain-community==0.2.16
langchain-openai==0.1.23
tavily-python>=0.3.0
unstructured
python-docx
openai
python-dotenv
pypdf
tiktoken
streamlit-pdf-viewer
fastapi
uvicorn[standard]
python-multipart
email-validator
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
pymupdf>=1.24.0
